Refactor: MapSection컴포넌트 분리 및 유지보수 향상
